third_party/WebKit/LayoutTests/fast/events/pointerevents/mouse-pointer-preventdefault-expected.txt

Verifies the effect of pointer event prevent-default on following pointer & mouse events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


--- test with preventDefault on none ---
- start with mouse outside target -
- move into target & click -
1. target received pointerover
2. target received pointerenter
3. target received mouseover
4. target received mouseenter
5. target received pointermove
6. target received mousemove
7. target received pointerdown
8. target received mousedown
9. target received pointerup
10. target received mouseup
- drag out of and into target & release within target -
11. target received pointerdown
12. target received mousedown
13. target received pointermove
14. target received mousemove
15. target received pointerout
16. target received pointerleave
17. target received mouseout
18. target received mouseleave
19. target received pointerover
20. target received pointerenter
21. target received mouseover
22. target received mouseenter
23. target received pointermove
24. target received mousemove
25. target received pointerup
26. target received mouseup
- move outside target again -
27. target received pointerout
28. target received pointerleave
29. target received mouseout
30. target received mouseleave

--- test with preventDefault on pointerdown ---
- start with mouse outside target -
- move into target & click -
1. target received pointerover
2. target received pointerenter
3. target received mouseover
4. target received mouseenter
5. target received pointermove
6. target received mousemove
7. target received pointerdown, prevented default
8. target received pointerup
- drag out of and into target & release within target -
9. target received pointerdown, prevented default
10. target received pointermove
11. target received pointerout
12. target received pointerleave
13. target received mouseout
14. target received mouseleave
15. target received pointerover
16. target received pointerenter
17. target received mouseover
18. target received mouseenter
19. target received pointermove
20. target received pointerup
- move outside target again -
21. target received pointerout
22. target received pointerleave
23. target received mouseout
24. target received mouseleave

--- test with preventDefault on pointerup ---
- start with mouse outside target -
- move into target & click -
1. target received pointerover
2. target received pointerenter
3. target received mouseover
4. target received mouseenter
5. target received pointermove
6. target received mousemove
7. target received pointerdown
8. target received mousedown
9. target received pointerup, prevented default
10. target received mouseup
- drag out of and into target & release within target -
11. target received pointerdown
12. target received mousedown
13. target received pointermove
14. target received mousemove
15. target received pointerout
16. target received pointerleave
17. target received mouseout
18. target received mouseleave
19. target received pointerover
20. target received pointerenter
21. target received mouseover
22. target received mouseenter
23. target received pointermove
24. target received mousemove
25. target received pointerup, prevented default
26. target received mouseup
- move outside target again -
27. target received pointerout
28. target received pointerleave
29. target received mouseout
30. target received mouseleave

--- test with preventDefault on pointerenter ---
- start with mouse outside target -
- move into target & click -
1. target received pointerover
2. target received pointerenter, prevented default
3. target received mouseover
4. target received mouseenter
5. target received pointermove
6. target received mousemove
7. target received pointerdown
8. target received mousedown
9. target received pointerup
10. target received mouseup
- drag out of and into target & release within target -
11. target received pointerdown
12. target received mousedown
13. target received pointermove
14. target received mousemove
15. target received pointerout
16. target received pointerleave
17. target received mouseout
18. target received mouseleave
19. target received pointerover
20. target received pointerenter, prevented default
21. target received mouseover
22. target received mouseenter
23. target received pointermove
24. target received mousemove
25. target received pointerup
26. target received mouseup
- move outside target again -
27. target received pointerout
28. target received pointerleave
29. target received mouseout
30. target received mouseleave

--- test with preventDefault on pointerleave ---
- start with mouse outside target -
- move into target & click -
1. target received pointerover
2. target received pointerenter
3. target received mouseover
4. target received mouseenter
5. target received pointermove
6. target received mousemove
7. target received pointerdown
8. target received mousedown
9. target received pointerup
10. target received mouseup
- drag out of and into target & release within target -
11. target received pointerdown
12. target received mousedown
13. target received pointermove
14. target received mousemove
15. target received pointerout
16. target received pointerleave, prevented default
17. target received mouseout
18. target received mouseleave
19. target received pointerover
20. target received pointerenter
21. target received mouseover
22. target received mouseenter
23. target received pointermove
24. target received mousemove
25. target received pointerup
26. target received mouseup
- move outside target again -
27. target received pointerout
28. target received pointerleave, prevented default
29. target received mouseout
30. target received mouseleave

--- test with preventDefault on pointerover ---
- start with mouse outside target -
- move into target & click -
1. target received pointerover, prevented default
2. target received pointerenter
3. target received mouseover
4. target received mouseenter
5. target received pointermove
6. target received mousemove
7. target received pointerdown
8. target received mousedown
9. target received pointerup
10. target received mouseup
- drag out of and into target & release within target -
11. target received pointerdown
12. target received mousedown
13. target received pointermove
14. target received mousemove
15. target received pointerout
16. target received pointerleave
17. target received mouseout
18. target received mouseleave
19. target received pointerover, prevented default
20. target received pointerenter
21. target received mouseover
22. target received mouseenter
23. target received pointermove
24. target received mousemove
25. target received pointerup
26. target received mouseup
- move outside target again -
27. target received pointerout
28. target received pointerleave
29. target received mouseout
30. target received mouseleave

--- test with preventDefault on pointerout ---
- start with mouse outside target -
- move into target & click -
1. target received pointerover
2. target received pointerenter
3. target received mouseover
4. target received mouseenter
5. target received pointermove
6. target received mousemove
7. target received pointerdown
8. target received mousedown
9. target received pointerup
10. target received mouseup
- drag out of and into target & release within target -
11. target received pointerdown
12. target received mousedown
13. target received pointermove
14. target received mousemove
15. target received pointerout, prevented default
16. target received pointerleave
17. target received mouseout
18. target received mouseleave
19. target received pointerover
20. target received pointerenter
21. target received mouseover
22. target received mouseenter
23. target received pointermove
24. target received mousemove
25. target received pointerup
26. target received mouseup
- move outside target again -
27. target received pointerout, prevented default
28. target received pointerleave
29. target received mouseout
30. target received mouseleave

--- test with preventDefault on pointermove ---
- start with mouse outside target -
- move into target & click -
1. target received pointerover
2. target received pointerenter
3. target received mouseover
4. target received mouseenter
5. target received pointermove, prevented default
6. target received mousemove
7. target received pointerdown
8. target received mousedown
9. target received pointerup
10. target received mouseup
- drag out of and into target & release within target -
11. target received pointerdown
12. target received mousedown
13. target received pointermove, prevented default
14. target received mousemove
15. target received pointerout
16. target received pointerleave
17. target received mouseout
18. target received mouseleave
19. target received pointerover
20. target received pointerenter
21. target received mouseover
22. target received mouseenter
23. target received pointermove, prevented default
24. target received mousemove
25. target received pointerup
26. target received mouseup
- move outside target again -
27. target received pointerout
28. target received pointerleave
29. target received mouseout
30. target received mouseleave

PASS successfullyParsed is true

TEST COMPLETE
